+++ src/animatedsprite.cpp
@@ -31,7 +31,7 @@ void AnimatedSprite::playAnimation(std::string animation, bool playOnce){
 	}
 }
 
-void AnimatedSprite::update(int elapsedTime){
+void AnimatedSprite::update(double elapsedTime){
 	timeElapsed += elapsedTime;
 
 	// if elapsed time exceeded set update time, update frame
